remove scikit learn testing module from normal flow
