+++ src/python/nimbusml/tests/scikit/test_uci_adult_scikit.py
@@ -16,7 +16,7 @@ from nimbusml.feature_extraction.categorical import OneHotVectorizer
 from nimbusml.linear_model import FastLinearBinaryClassifier
 from nimbusml.linear_model import LogisticRegressionClassifier
 from nimbusml.preprocessing.normalization import MeanVarianceScaler
-from nimbusml.utils import check_accuracy_scikit, get_X_y
+from nimbusml.utils import get_X_y
 from sklearn.base import clone
 from sklearn.datasets import load_iris
 from sklearn.decomposition import PCA
@@ -45,6 +45,19 @@ categorical_columns = [
     'native-country-region']
 selected_features = ['age', 'education-num']
 
+def check_accuracy_scikit(
+        test_file,
+        label_column,
+        predictions,
+        threshold,
+        sep=','):
+    (test, label) = get_X_y(test_file, label_column, sep=sep)
+    accuracy = np.mean(label[label_column].values == predictions.values)
+    assert_greater(
+        accuracy,
+        threshold,
+        "accuracy should be greater than %s" %
+        threshold)
 
 class TestUciAdultScikit(unittest.TestCase):
