+++ src/python/nimbusml/utils/__init__.py
@@ -1,5 +1,4 @@
-from .utils import get_X_y, evaluate_binary_classifier, check_accuracy, \
-    check_accuracy_scikit, load_img, ColumnSelector
+from .utils import get_X_y, evaluate_binary_classifier, load_img, ColumnSelector
 
 try:
     from inspect import signature
@@ -9,8 +8,6 @@ except ImportError:
 __all__ = [
     'get_X_y',
     'evaluate_binary_classifier',
-    'check_accuracy',
-    'check_accuracy_scikit',
     'load_img',
     'ColumnSelector',
     'signature'
